[search] Produce exact search matches everywhere
This CL simplifies the search code slightly by removing "inexact"
search matches (aka `SearchMatch`) alltogether. Inexact search matches
contain only a line number + the lines content, but not exactly how
many matches and where they are found. We added `SearchMatchExact` to
describe an exact (single) match, including the index
where in the line the match is found as well as the match length.

This CL removes `SearchMatch` in favor of `SearchMatchExact`, but at
the same time we rename `SearchMatchExact` to `SearchMatch` to reduce
the noise on the CL.

The simplification comes from not having 2 different kinds of matches
flowing through the various abstraction layers at the same time.

To produce exact search matches everywhere, we need to do two things:

  1) When searching in raw text directly we now use
     performExtendedSearchInContent instead of performSearchInContent.

  2) When we search in the backend, we immediately turn the backend
     search result into exact matches by finding all the matches
     in the backend search result.
     This comes with a slight overhead, but we'll be able to offset
     this by finally re-using this search result (follow-up CL).

Note that search tests for the various content providers are web tests
that we have disabled and will re-baseline once this CL has landed.

R=jarin@chromium.org

Bug: 1468875
Change-Id: If340217ac3eed5db3e640a2b213e817ba64c96ac