[test] Remove second TargetManager global
Before this CL, the "EnvironmentHelpers.ts" file would maintain a
separate TargetManager global, parallel to the TargetManager.instance().

This is iffy in unit tests: The first time a unit test calls
`createTarget`, we would nuke the existing TargetManager. Classes
already set-up would have used the wrong TargetManager.

An example is the DebuggerWorkspaceBinding. If the instance is created
with the default TargetManager.instance() but we call 'createTarget()'
later, the DebuggerWorkspaceBinding would not get notified about any
of the models since `createTarget` nuked the global.

To fix this, we rely solely on the TargetManager.instance() global.
EnvironmentHelpers.ts will still nuke the TargetManager.instance() in
"afterEach", so each test gets a fresh TargetManager. But otherwise
'createTarget' will no longer interfere.

R=jarin@chromium.org

Bug: None
Change-Id: Idf35f39e851122d71d4e284d101e857989d9fda7