[console] Disable assertion for console groups
The assertion was added with https://crrev.com/c/3364094 but seems not
to hold for many web-sites. The resulting terminal spam makes it very
noisy for chromium devs to work with local builds and the failed
assertion messages can even fill up syslog.

This CL removes the assertion in favor of a TODO.

R=jarin@chromium.org

Bug: 1477675
Change-Id: Ic3da8822b958f324ee2d77b627ecd296798098da