Support warnings in ThreadAppender in highlightedEntryInfo
These are shown as part of the info shown in a pop up when the user
hovers an event. These have been taken from TimelineUIUtils as is.

This CL adds the building of the warning to the track appender, so that
we can start getting rid of bits in TimelineUIUtils when we ship sync
tracks.

Took the opportunity to add special handling for ParseHTML events at
the same point, since these events contain are handled like so in the
legacy code.

Bug: 1478195
Change-Id: I6fb69ec5ab6eb3d8831dad6dfd3f71707918be1e