prerender: remove properties/capabilities of ResourceTreeModel and ResourceTreeFrame
The old prerendering section was removed in the following CL because
the same information is available in the new panel's "this page" pane.
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4835675

To fully remove the code related to the old prerendering section,
this CL removes properties/capabilities of ResourceTreeModel and
ResourceTreeFrame since preloading panels is built on top of
PreloadingModel.

Bug:1445219
Change-Id: I2e503a063a1eac0bff1a028acb073dde72107a59