Remove showNativeFunctionsInJSProfile setting.
This CL removes the setting and keeps the behaviour - enabling it by
default (which was true of the setting too - it was enabled by default).

Honestly, we aren't too sure what this setting does if you disable it,
and we cannot come up with any valid reasons to disable it. So we are
removing it. Some more context in the attached bug.

Fixed: 1479471
Change-Id: Ia0d6346dd8964c1eb9c8148e43c1dd48f0b79a3d