Update device emulation user agent strings
The default set of devices are now on years-old browsers, no longer reflecting the user agents sent by the majority of users.

This updates most of the top default devices.  I used BrowserStack heavily to confirm that the user agents reflect physical devices and what they return.  A few notable changes up for discussion:

1. Moved iOS devices to use Safari user agents instead of CriOS (Chrome iOS) based on popular usage.
2. Larger format iPads return desktop MacOS user agents, and I've updated those devices to reflect.
3. I added in the largest format iPhone and iPad.
4. https://developer.chrome.com/blog/user-agent-reduction-android-model-and-version/ indicates that some of the Android user agents should no longer be returning the updated Android version and device name in the UA string.  I didn't see this on BrowserStack but did see this on a personal Pixel 4a.  I'm not sure what the intended behavior is here but it may affect future UA updates.

I will specifically call out that this appears tto be the first case of "desktop" capabilities in this file and I would especially love eyes on that in case it isn't implemented as expected.

Change-Id: Id9bc431b06232e6df18b6eef36b5806c73c67490