Reland "RPP: Don't throw on StreamingCompileScriptWaiting events"
This is a reland of commit 831fe390cff883508e29d506ea1e0a40c72a9ca7

It contains a single change compared to the original: the unit tests
now asserts a hsl color instead of a CSS variable. This is needed
because of another revert (https://crrev.com/c/4853497).

Original change's description:
> RPP: Don't throw on StreamingCompileScriptWaiting events
>
> Some older logic wanted to dim the color of this event. When we got
> CSS custom props for the categories, the color parsing broke, and
> any traces with it fail to load.
>
> Bug: 1480205
> Change-Id: I5683de590c6ef77e62b5a1e1c3f7d0998bd4ccbe
> Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4851054
> Auto-Submit: Paul Irish <paulirish@chromium.org>
> Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
> Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>

Bug: 1480205
Change-Id: Idd2ced35f87ee01f3620e3eb66a87c3e521f3cfd