Support ignore listing in ThreadAppender
A relatively low hanging fruit. Ignore listing is applied to stacks with
callframes with the same URL. The rule is we "collapse" an ignored listed
stack into the bottom frame and add an entry with the "On ignore list"
title.

Because we need to keep track of each entry's parent, this CL uses the
node tree in the RendererHandler. Using this tree we can also determine
the flamechart level of each entry (instead of looking up for the first
free space in flame chart), something which this CL also does.

Bug: 1428024
Change-Id: I206298bf3f04c0ad600d45c85d2bcb6241a16137