Shorten filename embedded by code_generator_frontend.py
The code generator embeds the last n entries of its own filename into
the generated InspectorBackendCommands. It used to be n=4, which
happens to create a different filename depending on whether the
script is run from a chromium or a standalone devtools-frontend
checkout. So set n=3.

DISABLE_THIRD_PARTY_CHECK=Updating code genrator

Bug: none
Change-Id: I09fad6d43377c8558709cbe41dba106f9c813caf