Resolve sourcemaps in Perf Panel on primaryPageTarget
When we moved Tracing onto the root target we mistakenly updated this
code as well. But the root target does not implement the Debugger domain
so we were never able to resolve sourcemaps. We have to use the primary
page target for this.

However, there is still a scenario where:

1. Begin tracing with primaryPageTarget A (call this ppt-A)
2. ppt changes to ppt-B
3. Tracing is resolved

At this point, we have likely lost access to ppt-A and any sourcemaps
that were on that target. It is hard to know right now how common this
is and if this will cause issues. It won't break anything - the perf
panel will just silently show code without any sourcemap resolution -
but it might be something we need to revisit in the future. We could
maintain a store of all sourcemaps and keep them even as the
primaryPageTarget changes if we need to.

Bug: 1477281
Change-Id: I3c163eac774fc530e1500dba9a793590a9a0ecf5