Use a flag that lets us switch the source of sync tracks
The flag contains three values: old, new engine and both engines. The
default value, which is used everywhere except for the basic component
example, is the old engine. Whenever we are ready to switch we change
the default value to new engine.

The value of the flag determines if new engine is run fully and which
tracks (legacy or track appenders) are used in the flame chart.

Bug: 1428024
Change-Id: I750bcb1847b6981cb8d2f8de9cabef8c933a150a