Actually ignore instant events in samples integrator
Before this CL there was a discrepancy between the old and new engine in
how profile calls are built from samples and trace events. The
discrepancy came from the fact that instant events should be ignored
during this process, but they weren't always ignored in the new engine.
Instead, we checked if an event is an instant every time an event
starts, but not every time an event finishes. To make sure we ignore
instant events, they are ignored before the event start and end
callbacks are called.

Bug: 1449191
Change-Id: I40a1a1b4203db9b5225207eb95ef87134759f6ec