Sync tracks: resolve entry names from the ProfileNode
When we update the names (for example, if we get a sourcemap), we do so
on the ProfileNode, not on the synthetic trace event we create (which is
a good approach as we treat trace events as strictly immutable). That
means though that to use the new name we have to look up the profile
node.

I was first worried about the performance of this, but the lookup is 3
`Map.get()` calls, and we do not have to loop over any data, so I don't
expect a big impact.

Bug: 1475936
Change-Id: Id1b1a8a0375be9df738ee382e547fb7bf4ce7e51