Fix track filtering based on ThreadTracksSource flag
To enable the sync tracks migration to be easily toggled on/off, we
landed a flag that modifies which tracks are rendered onto the timeline.

However, this code was overly aggressive and if the setting was set to
OLD_ENGINE, it would remove other tracks such as Timings that have been
migrated to the new engine and should always be shown.

This CL updates the logic so that the behaviour is:

- For OLD_ENGINE
> Include all legacy tracks
> Include any TrackAppender tracks APART from ThreadAppender tracks

- For NEW_ENGINE
> Include no legacy tracks (note this does not include the Frames track
  which is managed separately and unaffected by this CL)
> Include all TrackAppender tracks (including ThreadAppender tracks)

Bug: 1484128
Change-Id: If42c2ce77ccc54c4549b7ba37793939bdc1d7fd0