Removed scrollbar on collapsed Network Track
There is a method in ChartViewPort than checks if the scrollbar should be shown at all times, however, it does not work for Network track because Network track has a fixed height and gets overlapped by other tracks. Therefore, here Flamechart informs ChartViewPort if the Network Track was collapsed/expanded.

Bug: 1481483
Change-Id: Ia5b2e8439817e87439ddcb5f5a66541aff067941