Collect screenshots in rbd results (artefacts)
Example failure with both image diff and screenshot generation only
https://luci-milo.appspot.com/ui/inv/u-liviurau-2023-09-25-12-57-00-5564594346cfc213/test-results?sortby=&groupby=

Note: since the summary HTML has a max length of 4000 we need to trim the error message for the resultdb.

A follow up is needed to publish all screenshots to resultdb and present them in Milo so we can get rid of all base64 encoded screenshots in the logs.

Bug: chromium:1484483
Change-Id: I12518b39223cb98121aa9b6aef56df7142b5e3bc