Fixed DevTools 'Learn more' link for Request Header Section
The DevTools 'Learn more' link does not meet the required standard ratio of 4.5:1. This change fixes that so the link color uses the variable "sys-color-primary" same as how devtools links are defined in other areas.  In addition, a fix for high contrast mode is also added.

1)
Before:
https://imgur.com/a/ptxTcue
After:
https://imgur.com/a/dKypw8y

2)
Before:
https://imgur.com/a/brwOQzO
After:
https://imgur.com/a/Mhhftir

Bug: 1486693
Change-Id: Ie050a1def79bba789ffc25d1111c8db61deef736