Defer extension loading until after navigation to a non-blocked host
When an extension is loaded while the target is a privileged (e.g.,
with chrome:// scheme) or blocked host (by runtime-blocked-hosts
policy), don't discard the extension entirely but defer loading until
after the next navigation.

Drive-bys: Fix a loading bug that prevented extension loads from being deferred twice, and fix a test that should have used a spy and not a stub.
Fixed: 1476264
Change-Id: If6b324faddb8500f8701edc0f8972830fd5d64d2