Add screenshot tests for ThreadTracksSource
This CL adds some screenshots that use either the old or new engine
rather than the BOTH flag, this makes sure we cannot cause a bug where
we are rendering the wrong set of tracks.

Once the migration is complete, we can remove the OLD_ENGINE flag and
then we will be able to remove these tests.

Fixed: 1484128
Change-Id: I7b7b1890404f847bd94a9f4b8fdd54ee60e4e2db