Adding <audio>/<video> link adorners
In an effort to increase the accessibility of media panels, this CL
adds a new link adorner on media elements inside the DOM Tree
of DevTool's Elements Panel. Upon clicking on this link adorner,
users will be directed to the media panel.

Notice: This CL is done by a student developer Junseo (Jeremy) Yoo
as part of Google Summer of Code 2023

Bug: 1448214
Change-Id: I7eabcc67359f883cac2e9a1406b233999cefedf7