Do not schedule track details if an event is selected
Debouncing the rendering of a track's details when an event in it has
been selected means the event's details will disappear and be replaced
with the track's details. Since we still want to debounce the rendering
in paths called repeatedly in short times (like scrolling or zooming),
let's skip the debouncing only when a new event is selected.

Fixed: 1489260
Change-Id: I4c21fea31376a45243eb3e2076a336e70ea4e8ab