Move DOMDebugger instrumentation breakpoints to EventBreakpoints
DOMDebugger#setInstrumentationBreakpoint is deprecated for a mirrored
version on the new EventBreakpoints domain. This CL moves the
instrumentation breakoints from the DOMDebugger model to the
EventBreakpoints model.

After this CL, the moved breakpoints will also work for workers,
not just for frame targets. A follow-up CL will clean-up the dead
code bits in DOMDebugger.

R=dsv@chromium.org

Bug: 1445175
Change-Id: Icd6b0a668abd4243059a20eb64630ea5a3413e4b