[DevTools] Remove timeline_model-legacy.ts
It seems there are no more references to the legacy globals upstream.

The remaining references were in front_end/Tests.js, in a test that
references variables that were removed 3+ years ago. I noticed this test is only referenced on a devtools_browser test that has been disabled for 7 years (https://codereview.chromium.org/2794223002). Since we have
coverage for the tested scenario in our newer suites, I opted for
deleting the test.

Bug: 1442410
Change-Id: I16cfaf42628f88ec922a1324c26e7ec599f284ee