Enable replacing CSS stylesheets on the fly after file change for watch script
This CL updates `generate_css_files` script and `watch` script to enable us to get CSS file changes instantly applied for non-legacy CSS files.

Drive by:
* Added an initial build step before starting to watch changes(1)
* Improved logs to include the time of the change.
* Updated `watch` listener to only get activated on `change` events.

(1) was something I was always missing starting this script. In addition to that, we need to re-build the CSS files with the hot reload script for the specified target and environment variables.

You can try the behavior by:
* Running `npm run watch` and starting the watch mode.
* Opening the DevTools.
* Updating a non-legacy CSS file and expect to see the CSS change instantly on the opened DevTools.

Bug: none for now
Change-Id: I2aa3ed1b36584d1ef011119d5092d12271fb3017