Export a tree as part of the SamplesHandler data
Along with built profile calls, a tree is exported, just as the
RendererHandler would. This tree is intended to be used when profiling
Node / JS only.

This includes an optimization where created profile calls don't need to
be sorted. This is achieved by leveraging the fact that the forEachFrame
helper calls back in hierarchical order, so we can store ProfileCalls in
the order in which the corresponding callback is done.

Note: to build the tree, this CL doesn't make use of the treify helper.
At first I thought we could, but doing so would imply going through all
profile calls to build the tree. Instead, we can build the tree as we
create profile calls.
Bug: 1462387
Change-Id: I05b4f4310aa2531d74476187a845400fd11d3982