Create new TraceBounds service
We have an issue in the codebase where the state of the active window
and the region the user is seeing in the timeline is not contained in
one single place and instead multiple components emit and listen to
multiple events in order to update.

To try to tackle this, this CL introduces a new TraceBounds service.
This service is instantiated when a trace is imported/recorded with the
entire trace bounds, and can then be used as the single source of truth
for any component.

This CL introduces the service and tests, and does instantiate it when a
trace is recorded by the user. It does not use it yet for any UI. The
rollout plan will be to first have all components that can change the
window also update the trace bounds service with the active window. Then
we can incrementally add listeners to the new service to have each UI
part moved one by one over to the new service, before removing the old
one.

Bug: 1491689
Change-Id: I338ae06319220113d2cfb72217c652b83db6f10f