Update TraceBounds when user takes action
This CL adds to all the callsites of performanceModel.setWindow() to
also call the BoundsManager to update the currently active bounds.

As a drive-by improvement, it also updates the function for zooming into
main thread activity to return the values rather than directly set them.
This reduces the amount of callsites we have to migrate and also makes
it easier when we need to migrate this method onto the new engine.

This CL does NOT use the BoundsManager for rendering any UI, that is
what I am working towards over this series of CLs.

Bug: 1491689
Change-Id: If088dd8d0189490143a0484427d79e928efa1078