[RPP] Use new engine in cpu profiling
Whenever we are in cpu profiling mode, we take the tree built by the samples handler and use it in the ThreadAppender

The caveat is this relies on the fake CpuProfile event injected by TimelineLoader, which is made available to the trace engine via the tracing model instance it uses. Whenever we want to remove the old engine, we need to figure out another approach to inject said event (or not inject any event at all).

Bug: 1462387
Change-Id: I4702205220fae122a4e6259eab8dd8b505881d11