Move localized breakpoint categories from sdk/ to ui/
This CL introduces a new `Category` enum that we use on the sdk/ layer`
instead of localized strings. Localization now happens on the ui/
layer.

Note that we preserve the existing behavior of sorting the categories
alphabetically given a local. We only slightly improve this by using
String.prototype.localeCompare instead of default string comparison.

R=dsv@chromium.org

Bug: 1445175
Change-Id: I28282a87a504355f48dc250dbdff6ce323035946