[docs] Streamline documentation.
Previously we had a lot of documentation bundled in the `workflows.md`
file, which wasn't necessarily obvious to new users. The getting the
code documentation now lives in a dedicated file `get_the_code.md`,
which is also directly linked from the toplevel `README.md`, since
that's what developers are looking first.

It also rearranges the rest of the project documentation a bit to move
the most important things to the top, add [TOC]s everywhere and
streamlines the documentation around (3rd party) dependencies, as well
as updating the documentation on testing.

Bug: none
Change-Id: I7b5d9121c7f1137eeff7a4ff159816908ef10705