[l10n] Move localized event breakpoint names out of sdk/
This CL moves the localized breakpoint names into a new file in
'panels/sources'. The localized names are required in two places:
The debugger paused message as well as the sidebar panel.

The CL only shuffles code around, the behavior stays the same and is
checked by existing unit tests.

R=dsv@chromium.org

Bug: 1445175
Change-Id: I48c1b572ed9a239a585d1449ac2b1f36c9b71214