[debugger] Mitigate race when resetting breakpoints.
In the scenario of code splitting, a single source location can map to
generated locations in various bundles that are loaded independently.
Whenever a new bundle is loaded, we check whether the back-end state
still matches what we think should be the state, and if not we first
throw away all the back-end breakpoints and then set the new positions.

This is however inherently racy without instrumentation breakpoints. So
to mitigate the issue, we don't check whether the new state exactly
matches the back-end state, but rather whether the new state contains
at least a single position AND is a proper subset of the back-end state.
In that case we do nothing, but just leave the back-end state as-is.

This way, when the back-end already has breakpoints set for the file in
multiple bundles, and we reload, we ensure that the front-end doesn't
delete the breakpoint from the unknown bundles when the first bundle is
loaded by the browser.

Fixed: chromium:1490369
Demo: https://devtools-source-identities.glitch.me/webpack-code-split/
Change-Id: I2dd29e31d82661544fbaaa6c332c3d3f7cdfdbd9