[Animations] Enable scrubbing by clicking on timeline header
Funnily, we've had the feature for seeking an animation group by
clicking on the timeline for a loooong time. However, at some point,
because of z-index issues, the click area has fallen back and became
non-clickable.

This CL updates it and improves the behavior by handling play behavior
after scrubbing an animation:
* It continues to play after scrubbing if it is already playing
* It doesn't play after scrubbing when it is stopped already.

Removed drag behavior for the playhead itself since we don't need to specifically drag the playhead for seeking an animation.

Fixed: 1492460
Change-Id: I2d524731982d96598e9df3ca4f1072e5032c1a62