[Coverage] Fix Coverage info is not populated on reload
Issue:
The Coverage tool does not display the Coverage result for the page reload, unless until there are more Coverage changes happen after, and the view then get updated.

Cause:
After then content loaded, a coverage was taken, but the Coverage view never got notified when the results returned.

Fix:
Notify the Coverage view to do an update after we get the results.

Bug: 1494457
Change-Id: Ia8ec7dc97478565d48451d75113c7b0ec41bcd3c