Only offer add to ignore list context menu if folder has unignored files
The context menu for a folder in the source tree view may have
menu items to add or remove that folder from the ignore list.
Previously we would give a remove option if there were rules
affecting all files in that folder that we could identify to
remove, and otherwise we provided options to add the folder to
the ignore list. However we gray out a folder if every item in
it is ignore listed. This is confusing in cases where everything
is ignored in a folder but due to rules that can't be connected
to the folder. With this change, we show no ignore listing options
if all files in a folder are ignored and we don't know what rule
to remove.

Bug: 1491823
Change-Id: Ie48fd24322765106da661902c0883c9970805bce