Add whiskers to interactions
This CL updates the interactions track to have whiskers to represent the
3 phases of an interaction.

|----------[=======]-------------|

=> The left whisker is the event's start time (event.ts)
=> The box start is the event's processingStart time
=> The box end is the event's processingEnd time
=> The right whisker is the event's end time (event.ts + event.dur)

It also updates the UI to show these numbers in the breakdown.

This CL also fixes a bug whereby when we were preventing nested
interactions we would often choose the event with no processing time;
this is incorrect when there is an event of the same total time range
but with processing time; that shows it is the event with the slow event
handler and as such is the one that we want to use.

There is more to do for this feature that will be done in follow-up CLs:

- We have lost the warning triangle now that we do not draw a box up to
  the end of the entry. I will send a follow-up CL that adds that back
  to the box.
- We want to show a breakdown of the numbers in a tooltip. I will do
  that in a follow-up CL also.

Bug: 1495751
Change-Id: I7be0a0d9858d798099577143ca2f1e3da598528a