Show cookie deprecation metadata (allowlist) issues in the DevTools frontend panel.
This was implemented in https://chromium-review.googlesource.com/c/chromium/src/+/4916620 as a new Issue type. In the long run we will make it a Warning type of CookieIssue. This will involve more refactoring in the future, but we will store the description in "cookieWarnMetadataGrant[Read/Set].md" for this future pattern.

Bug: b/305738703
Change-Id: I76b348352e77d0012a221f192f04d2f2aafc8896