INP: maximise processing time when hiding child events
When we pick the interaction to not nest, we might under-represent its
processing time if one of the nested children had a logner processing
time. To avoid this, we look through all the child events and pick the
earlierst processing start and latest processing end to show a fair
reflection.

Bug: 1495751
Change-Id: I9f050eafb45b09c876edebe8823a065927dd39ef
Co-authored-by: paulirish@chromium.org